Delegate all time intervals creation to DomainCollection

+++ src/calendar/DomainCollection.js
@@ -1,21 +1,23 @@
 export default class DomainCollection {
-  constructor(args) {
-    this.collection = new Map(args);
+  constructor(dateHelper, interval, start, range) {
+    this.collection = new Map();
+
+    if (interval && start && range) {
+      this.collection = new Map(
+        dateHelper
+          .intervals(interval, start, range)
+          .map((d) => (Array.isArray(d) ? d : [d])),
+      );
+    }
+
     this.min = null;
     this.max = null;
     this.keys = [];
+    this.dateHelper = dateHelper;
 
-    if (args) {
-      this.#resetKeys();
+    if (this.collection.size > 0) {
+      this.#refreshKeys();
     }
-  }
-
-  setBatch(keys) {
-    keys.forEach((key) => {
-      this.collection.set(key);
-    });
-
-    this.#resetKeys();
   }
 
   has(key) {
@@ -26,51 +28,74 @@ export default class DomainCollection {
     return this.collection.get(key);
   }
 
-  set(key, value) {
-    this.collection.set(key, value);
-    this.#resetKeys();
-
-    return this.has(key);
-  }
-
   delete(key) {
     const r = this.collection.delete(key);
-    this.#resetKeys();
+    this.#refreshKeys();
 
     return r;
   }
 
   forEach(callback) {
-    return this.collection.forEach((element) => callback(element));
+    return this.collection.forEach(callback);
+  }
+
+  at(index) {
+    return this.keys[index];
   }
 
   keyIndex(d) {
     return this.keys.indexOf(d);
   }
 
-  pop() {
-    const r = this.collection.delete(this.max);
-    this.#resetKeys();
-
-    return r;
-  }
-
-  shift() {
-    const r = this.collection.delete(this.min);
-    this.#resetKeys();
-
-    return r;
-  }
-
-  trim(count, fromBeginning) {
-    if (this.keys.length > count) {
-      return fromBeginning ? this.shift() : this.pop();
+  clamp(minDate, maxDate) {
+    if (minDate && this.min < minDate) {
+      this.keys.filter((key) => key < minDate).forEach((d) => this.delete(d));
     }
 
-    return false;
+    if (maxDate && this.max > maxDate) {
+      this.keys.filter((key) => key > maxDate).forEach((d) => this.delete(d));
+    }
+
+    return this;
   }
 
-  #resetKeys() {
+  merge(newCollection, limit, subDomainCallback) {
+    newCollection.keys.forEach((domain, index) => {
+      if (this.has(domain)) {
+        return;
+      }
+
+      if (this.collection.size >= limit) {
+        if (domain > this.max) {
+          this.collection.delete(this.min);
+        } else {
+          this.collection.delete(this.max);
+        }
+      }
+      this.collection.set(domain, subDomainCallback(domain, index));
+      this.#refreshKeys();
+    });
+  }
+
+  slice(limit, fromBeginning = true) {
+    if (this.keys.length > limit) {
+      const keysToDelete = fromBeginning ?
+        this.keys.slice(0, -limit) :
+        this.keys.slice(limit);
+
+      keysToDelete.forEach((key) => {
+        this.delete(key);
+      });
+    }
+
+    return this;
+  }
+
+  debug() {
+    console.log(this.keys.map((d) => new Date(d)));
+  }
+
+  #refreshKeys() {
     this.keys = Array.from(this.collection.keys())
       .map((d) => parseInt(d, 10))
       .sort((a, b) => a - b);
@@ -79,5 +104,7 @@ export default class DomainCollection {
     // eslint-disable-next-line prefer-destructuring
     this.min = keys[0];
     this.max = keys[keys.length - 1];
+
+    return this.keys;
   }
 }
